move LandingPage directly to root page.tsx
